backend/requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0   # PostgreSQL driver

# ── Google OAuth ─────────────────────────────────────────────────────────────
google-auth>=2.28.0       # id_token.verify_oauth2_token

# ── File Upload & Multipart ──────────────────────────────────────────────────
python-multipart>=0.0.9   # Required for FastAPI UploadFile

# ── Data Validation ──────────────────────────────────────────────────────────
pydantic[email]>=2.6.0    # EmailStr used in schemas

# ── OCR (REST-based — no heavy local model) ──────────────────────────────────
requests>=2.31.0           # Google Cloud Vision REST API calls

# ── Environment Variables ────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0             # AsyncClient for FastAPI test client
